Add non-blocking AFE calibration method
Add an asynchronous (non-blocking) method for doing AFE calibration.
This allows the system to do other tasks while waiting for AFE
calibration to complete.

Also adds a paramenter to begin() to allow for skipping the
initializaion and calibration steps.  This allows a user to manually
initialize the ADC with custom configuration and then calibrate it
just once.